fix: clean up route export warnings
Move testable page components out of TanStack route files so file-based route modules export only Route.

Update route page tests to import the moved components from feature modules, keep route wiring coverage on Route.options, and document the split route testing convention.

Stabilize the custom collection deletion e2e by reloading Home before asserting persisted custom collection state.

Closes #69
