+++ src/components/features/DictionaryPage.tsx
@@ -1,0 +1,5 @@
+import { DictionaryBrowser } from '@/components/features/DictionaryBrowser'
+
+export function DictionaryPage() {
+    return <DictionaryBrowser />
+}
